// src/pages/index/interfaces.ts
export interface Configuration {
  numberOfYears: number;
  totalAssets: number;
  annualizedRate: number;
  income: number;
  incomeIncreaseRate: number;
  expense: number;
  inflationRate: number;
  maximumIncome: number;
  minimumIncome: number;
  minimumIncomeIncreaseRate: number;
  workingYears: number;
}